Updated Class Diagram as of November 25, 2025
In this class diagram, we have visually illustrated the connection of each class and how different fields are utilized to streamline the order process. We use this class diagram to also understand the design patterns that we can use in our code. Still may be updated in future if we plan to expand our system.